Fixed Windows cross compilation
The cross compilation Dockerfile for windows was breaking for me when I
ran it, on a heredoc error. This fixes that, and then injects the
config.toml file we previously had in the repo into the Dockerfile
directly. This ensures that we can cross compile for Windows
successfully, without breaking native Windows builds.