Revert sounds back to signed
I previously refactored sprites to be unsigned because they will never
be negative. In that refactor, I incorrectly carried that through for
sound data, because sv_special handles *both* sprite data (which is
unsigned) and volume/pan data (which is signed). Reverted to old
behavior, with explicit casts to ints when calling play_sound.